Support code block part in inline chat (#195725)
* extract CodeBlockPart

* separate css

* layout updating

* Use code block actions in inline chat

* improve css

* polish

* fix test mocks
