+++ src/vs/workbench/contrib/chat/browser/codeBlockPart.css
@@ -1,0 +1,55 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+
+.interactive-result-code-block {
+	position: relative;
+}
+
+.interactive-result-code-block .monaco-toolbar {
+	display: none;
+	position: absolute;
+	top: -13px;
+	right: 10px;
+	height: 26px;
+	background-color: var(--vscode-interactive-result-editor-background-color, var(--vscode-editor-background));
+	border: 1px solid var(--vscode-chat-requestBorder);
+	z-index: 100;
+}
+
+.interactive-result-code-block .monaco-toolbar .action-item {
+	height: 24px;
+	width: 24px;
+	margin: 1px 2px;
+}
+
+.interactive-result-code-block .monaco-toolbar .action-item .codicon {
+	margin: 1px;
+}
+
+.interactive-result-code-block:hover .monaco-toolbar,
+.interactive-result-code-block .monaco-toolbar:focus-within,
+.interactive-result-code-block.focused .monaco-toolbar {
+	display: initial;
+	border-radius: 2px;
+}
+
+.interactive-result-code-block {
+	margin: 16px 0;
+}
+
+.interactive-result-code-block .interactive-result-editor .monaco-editor {
+	border: 1px solid var(--vscode-input-border, transparent);
+}
+
+.interactive-result-code-block .interactive-result-editor .monaco-editor.focused {
+	border-color: var(--vscode-focusBorder, transparent);
+}
+
+.interactive-result-code-block,
+.interactive-result-code-block .monaco-editor,
+.interactive-result-code-block .monaco-editor .overflow-guard {
+	border-radius: 4px;
+}
