+++ src/vs/workbench/contrib/chat/browser/media/chat.css
@@ -257,65 +257,16 @@
 	color: var(--vscode-icon-foreground) !important;
 }
 
-.interactive-item-container .interactive-result-editor-wrapper {
-	position: relative;
-}
-
-.interactive-item-container .interactive-result-editor-wrapper .monaco-toolbar {
-	display: none;
-	position: absolute;
-	top: -13px;
-	right: 10px;
-	height: 26px;
-	background-color: var(--vscode-interactive-result-editor-background-color, var(--vscode-editor-background));
-	border: 1px solid var(--vscode-chat-requestBorder);
-	z-index: 100;
-}
-
-.interactive-item-container .interactive-result-editor-wrapper .monaco-toolbar .action-item {
-	height: 24px;
-	width: 24px;
-	margin: 1px 2px;
-}
-
-.interactive-item-container .interactive-result-editor-wrapper .monaco-toolbar .action-item .codicon {
-	margin: 1px;
-}
-
-.interactive-item-container .interactive-result-editor-wrapper:hover .monaco-toolbar,
-.interactive-item-container .interactive-result-editor-wrapper .monaco-toolbar:focus-within,
-.interactive-item-container .interactive-result-editor-wrapper.focused .monaco-toolbar {
-	display: initial;
-	border-radius: 2px;
-}
-
-.interactive-item-container .interactive-result-editor-wrapper {
-	margin: 16px 0;
-}
-
-.interactive-session .interactive-item-container.interactive-response .interactive-result-editor-wrapper .interactive-result-editor .monaco-editor,
-.interactive-session .interactive-item-container.interactive-response .interactive-result-editor-wrapper .interactive-result-editor .monaco-editor .margin,
-.interactive-session .interactive-item-container.interactive-response .interactive-result-editor-wrapper .interactive-result-editor .monaco-editor .monaco-editor-background {
+.interactive-response .interactive-result-code-block .interactive-result-editor .monaco-editor,
+.interactive-response .interactive-result-code-block .interactive-result-editor .monaco-editor .margin,
+.interactive-response .interactive-result-code-block .interactive-result-editor .monaco-editor .monaco-editor-background {
 	background-color: var(--vscode-interactive-result-editor-background-color);
 }
 
-.interactive-item-container .interactive-result-editor-wrapper .interactive-result-editor .monaco-editor {
-	border: 1px solid var(--vscode-input-border, transparent);
-}
-
-.interactive-item-container .interactive-result-editor-wrapper .interactive-result-editor .monaco-editor.focused {
-	border-color: var(--vscode-focusBorder, transparent);
-}
-
-.interactive-item-container .interactive-result-editor-wrapper,
-.interactive-item-container .interactive-result-editor-wrapper .monaco-editor,
-.interactive-item-container .interactive-result-editor-wrapper .monaco-editor .overflow-guard {
-	border-radius: 4px;
-}
-
-.interactive-item-container.interactive-item-compact .interactive-result-editor-wrapper {
+.interactive-item-compact .interactive-result-code-block {
 	margin: 0 0 8px 0;
 }
+
 
 .interactive-response .interactive-response-error-details {
 	display: flex;
